+++ R/helper-allelic.R
@@ -5,10 +5,13 @@
 #' are marked with the following suffix: an underscore and
 #' a consistent symbol for each of the two alleles,
 #' e.g. \code{ENST123_M} and \code{ENST123_P},
-#' or \code{ENST123_alt} and \code{ENST123_ref}.
-#' There must be exactly two alleles for each transcript, 
+#' or \code{ENST123_alt} and \code{ENST123_ref}, etc.
+#' \code{importAllelicCounts} requires the tximeta package.
+#' Further information in Details below.
+#'
+#' \strong{Requirements} - There must be exactly two alleles for each transcript, 
 #' and the \code{--keep-duplicates} option should be used in
-#' Salmon indexing to avoid removing transcripts with identical sequence.
+#' Salmon indexing to avoid removal of transcripts with identical sequence.
 #' The output object has half the number of transcripts,
 #' with the two alleles either stored in a \code{"wide"} object,
 #' or as re-named \code{"assays"}. Note carefully that the symbol
@@ -17,13 +20,27 @@
 #' (see the \code{format} argument description and Value
 #' description below).
 #'
-#' Requires the tximeta package.
+#' \strong{tx2gene} - The two columns should include the \code{a1} and \code{a2}
+#' suffix for the transcripts and genes/groups, or those will be added internally,
+#' if it is detected that the first transcript does not have these suffices.
+#' For example if \code{_alt} or \code{_ref}, or \code{_M} or \code{_P}
+#' (as indicated by the \code{a1} and \code{a2} arguments) are not present
+#' in the table, the table rows will be duplicated with those suffices
+#' added on behalf of the user.
+#' If \code{tx2gene} is not provided, the output object will be transcript-level.
+#' Note: do not attempt to set the \code{txOut} argument, it will
+#' conflict with internal calls to downstream functions.
+#' Note: if the a1/a2 suffices are not at the end of the transcript name
+#' in the quantification files, e.g. \code{ENST123_M|<metadata>},
+#' then \code{ignoreAfterBar=TRUE} can be used to match regardless of
+#' the string following \code{|} in the quantification files.
+#'
 #' \code{skipMeta=TRUE} is used, as it is assumed
 #' the diploid transcriptome does not match any reference
 #' transcript collection. This may change in future iterations
 #' of the function, depending on developments in upstream
 #' software.
-#'
+#' 
 #' @param coldata a data.frame as used in \code{tximeta}
 #' @param a1 the symbol for the effect allele
 #' @param a2 the symbol for the non-effect allele
@@ -38,22 +55,11 @@
 #' either \code{a1-} or \code{a2-}.
 #' @param tx2gene optional, a data.frame with first column indicating
 #' transcripts, second column indicating genes (or any other transcript
-#' grouping). Either this should include the \code{a1} and \code{a2}
-#' suffix for the transcripts and genes, or those will be added internally,
-#' if it is detected that the first transcript does not have these suffices.
-#' For example if \code{_alt} or \code{_ref}, or \code{_M} or \code{_P}
-#' (as indicated by the \code{a1} and \code{a2} arguments) are not present
-#' in the table, the table rows will be duplicated with those suffices
-#' added on behalf of the user.
-#' If not provided, the output object will be transcript-level.
-#' Note: do not attempt to set the \code{txOut} argument, it will
-#' conflict with internal calls to downstream functions.
-#' Note: if the a1/a2 suffices are not at the end of the transcript name
-#' in the quantification files, e.g. \code{ENST123_M|<metadata>},
-#' then \code{ignoreAfterBar=TRUE} can be used to match regardless of
-#' the string following \code{|} in the quantification files.
+#' grouping). Alternatively, this can be a GRanges object with columns
+#' \code{tx_id}, and \code{group_id} (see \code{makeTx2Tss}). For
+#' more information on this argument see Details.
 #' @param ... any arguments to pass to tximeta
-#'
+#' 
 #' @return a SummarizedExperiment, with allele counts (and other data)
 #' combined into a wide matrix \code{[a2 | a1]}, or as assays (a1, then a2).
 #' The original strings associated with a1 and a2 are stored in the
@@ -73,10 +79,20 @@ importAllelicCounts <- function(coldata, a1, a2,
   a1match <- paste0("_",a1,"$")
   a2match <- paste0("_",a2,"$")
 
-  txOut <- is.null(tx2gene) # output transcripts if no tx2gene provided  
+  txOut <- is.null(tx2gene) # output transcripts if no tx2gene provided
+  t2gRanges <- is(tx2gene, "GRanges")
   if (!txOut) {
+    # if tx2gene is ranges, then pull out the table for collapsing
+    if (t2gRanges) {
+      cols <- c("tx_id","group_id")
+      stopifnot(all(cols %in% names(mcols(tx2gene))))
+      # swap around variable names to run the data.frame-based code
+      txps <- tx2gene
+      tx2gene <- mcols(txps)[,cols]
+    }
     stopifnot(ncol(tx2gene) == 2)
     # see if tx2gene already has tagged the txps and genes
+    diploid <- FALSE # save whether tx2gene is diploid
     if (grepl(a1match, tx2gene[1,1]) | grepl(a2match, tx2gene[1,1])) {
       # ensure same number of a1 and a2 alleles in txps and genes
       sum1txp <- sum(grepl(a1match, tx2gene[,1]))
@@ -86,6 +102,7 @@ importAllelicCounts <- function(coldata, a1, a2,
       stopifnot(sum1txp > 0 & sum2txp > 0 & sum1gene > 0 & sum2gene > 0)
       stopifnot(sum1txp == sum2txp)
       stopifnot(sum1gene == sum2gene)
+      diploid <- TRUE
     } else {
       a2a1_vec <- rep(c(a2, a1), each=nrow(tx2gene))
       tx2gene <- data.frame(
@@ -94,7 +111,7 @@ importAllelicCounts <- function(coldata, a1, a2,
       )
     }
   }
-  
+
   se <- tximeta::tximeta(coldata, skipMeta=TRUE, tx2gene=tx2gene, txOut=txOut, ...)
 
   # remove any characters after "|"
@@ -128,10 +145,9 @@ importAllelicCounts <- function(coldata, a1, a2,
       a_wide
     })
     # make a new SE
-    wide <- SummarizedExperiment(assays=assays_wide,
-                                 colData=coldata_wide)
-    metadata(wide) <- c(metadata(se), list(alleles=c(a1=a1, a2=a2)))
-    return(wide)
+    out <- SummarizedExperiment(assays=assays_wide,
+                                colData=coldata_wide)
+    metadata(out) <- c(metadata(se), list(alleles=c(a1=a1, a2=a2)))
   } else if (format == "assays") {
     se_a1 <- se[txp_nms_a1,]
     se_a2 <- se[txp_nms_a2,]
@@ -143,8 +159,25 @@ importAllelicCounts <- function(coldata, a1, a2,
     # add the a2 matrices to the a1 SE object
     assays(se_a1) <- c(assays(se_a1), assays(se_a2))
     metadata(se_a1) <- c(metadata(se_a1), list(alleles=c(a1=a1, a2=a2)))
-    return(se_a1)
+    out <- se_a1
   }
+  if (t2gRanges) {
+    if (diploid) {
+      txps <- txps[grepl(a1match, names(txps))]
+      names(txps) <- sub(a1match,"",names(txps))
+      mcols(txps)$tx_id <- sub(a1match,"",mcols(txps)$tx_id)
+      mcols(txps)$group_id <- sub(a1match,"",mcols(txps)$group_id)
+    }
+    stopifnot(all(rownames(out) %in% mcols(txps)$group_id))
+    tx_list <- CharacterList(split(mcols(txps)$tx_id, mcols(txps)$group_id))
+    # TODO reduce with plyranges
+    txps <- txps[!duplicated(mcols(txps)$group_id)]
+    names(txps) <- mcols(txps)$group_id
+    txps <- txps[rownames(out)]
+    mcols(txps)$tx_id <- tx_list[rownames(out)]
+     SummarizedExperiment::rowRanges(out) <- txps
+  }
+  out
 }
 
 #' Make a GRanges linking transcripts to TSS within gene