+++ tests/testthat/test_compress.R
@@ -4,8 +4,9 @@ library(fishpond)
 
 test_that("compressing uncertainty works", {
 
+  m <- 100
   set.seed(1)
-  y <- makeSimSwishData(m=200, n=100)
+  y <- makeSimSwishData(m=m, n=100)
   y$batch <- factor(rep(3:1,c(30,40,30)))
   plotInfReps(y, 5, x="condition", reorder=TRUE)
   dev.off()
@@ -13,16 +14,13 @@ test_that("compressing uncertainty works", {
   dev.off()
 
   set.seed(1)
-  y <- makeSimSwishData(m=200, n=100, meanVariance=TRUE)
+  y <- makeSimSwishData(m=m, n=100, meanVariance=TRUE)
   y$batch <- factor(rep(3:1,c(30,40,30)))
   plotInfReps(y, idx=5, x="condition", reorder=TRUE)
   dev.off()
+  plotInfReps(y, idx=5, x="condition", reorder=TRUE, useMean=FALSE)
+  dev.off()  
   plotInfReps(y, idx=5, x="condition", cov="batch", reorder=TRUE)
-  dev.off()
-
-  y0 <- y
-  assays(y0) <- assays(y0)[1:3]
-  plotInfReps(y0, idx=5, x="condition", useMean=FALSE)
   dev.off()
   
   y2 <- y