+++ tests/testthat/test_correlation.R
@@ -7,7 +7,8 @@ test_that("swish can detect correlations with log counts", {
   set.seed(1)
 
   n <- 20
-  y <- makeSimSwishData(m=500, n=n, null=TRUE)
+  m <- 250
+  y <- makeSimSwishData(m=m, n=n, null=TRUE)
   nms <- c("counts",paste0("infRep",1:20))
   lambda1 <- exp(seq(4, 5, length.out=n))
   lambda2 <- rev(lambda1)
@@ -38,7 +39,7 @@ test_that("swish can detect correlations with log counts", {
   set.seed(5)
 
   n <- 20
-  y <- makeSimSwishData(m=500, n=n, null=TRUE)
+  y <- makeSimSwishData(m=m, n=n, null=TRUE)
   nms <- c("counts",paste0("infRep",1:20))
   lambda1 <- exp(seq(4, 5, length.out=n/2))
   lambda2 <- rev(lambda1)
@@ -69,15 +70,17 @@ test_that("swish can detect correlations with log counts", {
   #plot(-log10(mcols(y)$pvalue[1:30]))
   mcols(y)[1:4,]
   
-  #plotInfReps(y, 1, x="cov", cov="condition",
-  #            legend=TRUE, legendPos="top")
-
+  plotInfReps(y, 1, x="cov", cov="condition",
+              shiftX=.0005,
+              legend=TRUE, legendPos="top")
+  dev.off()
+  
   ### up-down-up pattern ###
 
   set.seed(1)
 
   n <- 40
-  y <- makeSimSwishData(m=500, n=n, null=TRUE)
+  y <- makeSimSwishData(m=m, n=n, null=TRUE)
   nms <- c("counts",paste0("infRep",1:20))
   cov <- rep(1:(n/4),each=2)
   lambda1 <- (cov - 1)*(cov - 10)*(cov - 10) + 100
@@ -108,7 +111,9 @@ test_that("swish can detect correlations with log counts", {
   #plot(-log10(mcols(y)$pvalue[1:30]))
   mcols(y)[1:4,]
   
-  #plotInfReps(y, 1, x="cov", cov="condition",
-  #            legend=TRUE, legendPos="top")  
+  plotInfReps(y, 1, x="cov", cov="condition",
+              shiftX=.05,
+              legend=TRUE, legendPos="bottom")
+  dev.off()
   
 })